Move index commands to subcommands
