+++ osu.ElasticIndexer/Commands/Index/CloseIndexCommand.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading;
 using McMaster.Extensions.CommandLineUtils;
 
-namespace osu.ElasticIndexer.Commands
+namespace osu.ElasticIndexer.Commands.Index
 {
     [Command("close", Description = "Closes unused indices.")]
     public class CloseIndexCommand : ProcessorCommandBase
