+++ osu.ElasticIndexer/Commands/Index/ListIndicesCommand.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using Elasticsearch.Net;
 using McMaster.Extensions.CommandLineUtils;
 
-namespace osu.ElasticIndexer.Commands
+namespace osu.ElasticIndexer.Commands.Index
 {
     [Command("list", Description = "Lists indices.")]
     public class ListIndicesCommand : ProcessorCommandBase
