+++ osu.ElasticIndexer/Commands/Index/UpdateAliasCommand.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading;
 using McMaster.Extensions.CommandLineUtils;
 
-namespace osu.ElasticIndexer.Commands
+namespace osu.ElasticIndexer.Commands.Index
 {
     [Command("alias", Description = "Updates alias to the latest index of a given version.")]
     public class UpdateAliasCommand
